Move Marshal to rely less on shared examples
